refactor(#10): move detect_python_bin to supervisor/_helpers.py
Pull the helper out of supervisor.py into a tiny stdlib-only module so
both supervisor.py and migrate.py import from one canonical location
instead of one importing the other. supervisor.py re-exports the symbol
for the existing test suite.
